requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.19.0
pycryptodome>=3.15.0
python-dotenv>=1.0.0